Complete overhaul of how translations and internationalization is handled in the framework.  All textdomains should be hardcoded strings.  Therefore, the previous method of using a function across both the framework and themes built from it won't work.  See: http://markjaquith.wordpress.com/2011/10/06/translating-wordpress-plugins-and-themes-dont-get-clever/
The framework and its extensions now use unique textdomains.  However, the framework filters 'gettext' to load theme translations to handle text from the framework and extensions.  This method will only work is the theme appropriately uses the template directory as its textdomain.  It's a bit of a compromise, but hooks do exist for themes to correct any changes they want to make.

Added the i18n.php file to separate language functions.  Added a few functions to facilitate this major change.

git-svn-id: http://svn.locallylost.com/themes/hybrid-core/trunk@900 dba0f204-706d-4bc1-bc29-8b92e0485636